requirements.txt
# Core dependencies
evrmore-rpc>=0.1.0
sqlalchemy>=1.4.0
alembic>=1.7.5
psycopg2-binary>=2.9.3
pyjwt>=2.3.0
cryptography>=36.0.0
passlib>=1.7.4

# API dependencies
fastapi>=0.75.0
uvicorn>=0.17.0
python-dotenv>=0.19.0
pydantic>=1.9.0
email-validator>=1.1.3
python-multipart>=0.0.5
requests>=2.27.0

# For development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.1.0
isort>=5.10.1 